Final Touch;

+++ src/main/java/com/raksha/response/InsurancesPlanResponse.java
@@ -2,47 +2,47 @@ package com.raksha.response;
 
 public class InsurancesPlanResponse {
 	
-	private double compreYear1;
-	private double compreYear2;
-	private double compreYear3;
-	private double thirdYear1;
-	private double thirdYear2;
-	private double thirdYear3;
+	private int compreYear1;
+	private int compreYear2;
+	private int compreYear3;
+	private int thirdYear1;
+	private int thirdYear2;
+	private int thirdYear3;
 	
-	public double getCompreYear1() {
+	public int getCompreYear1() {
 		return compreYear1;
 	}
-	public void setCompreYear1(double compreYear1) {
+	public void setCompreYear1(int compreYear1) {
 		this.compreYear1 = compreYear1;
 	}
-	public double getCompreYear2() {
+	public int getCompreYear2() {
 		return compreYear2;
 	}
-	public void setCompreYear2(double compreYear2) {
+	public void setCompreYear2(int compreYear2) {
 		this.compreYear2 = compreYear2;
 	}
-	public double getCompreYear3() {
+	public int getCompreYear3() {
 		return compreYear3;
 	}
-	public void setCompreYear3(double compreYear3) {
+	public void setCompreYear3(int compreYear3) {
 		this.compreYear3 = compreYear3;
 	}
-	public double getThirdYear1() {
+	public int getThirdYear1() {
 		return thirdYear1;
 	}
-	public void setThirdYear1(double thirdYear1) {
+	public void setThirdYear1(int thirdYear1) {
 		this.thirdYear1 = thirdYear1;
 	}
-	public double getThirdYear2() {
+	public int getThirdYear2() {
 		return thirdYear2;
 	}
-	public void setThirdYear2(double thirdYear2) {
+	public void setThirdYear2(int thirdYear2) {
 		this.thirdYear2 = thirdYear2;
 	}
-	public double getThirdYear3() {
+	public int getThirdYear3() {
 		return thirdYear3;
 	}
-	public void setThirdYear3(double thirdYear3) {
+	public void setThirdYear3(int thirdYear3) {
 		this.thirdYear3 = thirdYear3;
 	}
